Add files via upload
This project utilizes dijkstra's algorithm to find the optimal path from one location to another on UIC's east campus. The program finds the nodes of the chosen locations, along with their coordinates. It then takes the nodes of the closet buildings, builds a path and calculates the total miles to get to the location.